remove clickBreadcrumbs, use existing customEvents, skip some flakey tests on firefox
