--- packages/browser-integration-tests/suites/replay/clickBreadcrumbs/template.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-  </head>
-  <body>
-    <div role="button" id="error" class="btn btn-error" aria-label="An Error">An Error</div>
-    <button>
-      <img id="img"
-        alt="Alt Text"
-      />
-    </button>
-    <button class="sentry-unmask" aria-label="Unmasked label">
-      Unmasked
-    </button>
-  </body>
-</html>
