+++ packages/browser-integration-tests/suites/replay/customEvents/template.html
@@ -4,6 +4,14 @@
     <meta charset="utf-8" />
   </head>
   <body>
-    <button id="go-background">New Tab</button>
+    <div role="button" id="error" class="btn btn-error" aria-label="An Error">An Error</div>
+    <button>
+      <img id="img"
+        alt="Alt Text"
+      />
+    </button>
+    <button class="sentry-unmask" aria-label="Unmasked label">
+      Unmasked
+    </button>
   </body>
 </html>
